requirements.txt
arrow==1.2.2
atomicwrites==1.4.0
attrs==21.4.0
binaryornot==0.4.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
cookiecutter==1.7.3
coverage==4.5.4
idna==3.3
Jinja2==3.1.1
jinja2-time==0.2.0
MarkupSafe==2.1.1
more-itertools==8.12.0
packaging==21.3
pluggy==0.13.1
poyo==0.5.0
py==1.11.0
pyparsing==3.0.7
pytest==5.1.1
pytest-cov==2.7.1
python-dateutil==2.8.2
python-dotenv==0.10.3
python-slugify==6.1.1
PyYAML==6.0
requests==2.27.1
selenium==3.141.0
six==1.16.0
text-unidecode==1.3
urllib3==1.26.9
wcwidth==0.2.5
webdriver-manager==3.5.4
